+++ tools/nut-scanner/scan_nut.c
@@ -837,7 +837,7 @@ nutscan_device_t * nutscan_scan_ip_range_nut_authconf(nutscan_ip_range_list_t * 
 				" pass=%u",
 				__func__, max_threads_scantype,
 				curr_threads, thread_count,
-				pass
+				(unsigned int)pass
 			);
 		}
 # else
