+++ tools/nut-scanner/scan_xml_http.c
@@ -623,7 +623,7 @@ nutscan_device_t * nutscan_scan_ip_range_xml_http(nutscan_ip_range_list_t * irl,
 					" pass=%u",
 					__func__, max_threads_scantype,
 					curr_threads, thread_count,
-					pass
+					(unsigned int)pass
 				);
 			}
 # else
